// Front/src/api/generated/model/adminLabDto.ts
/**
 * Generated by orval v8.11.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { AdminTaskDto } from './adminTaskDto';

export interface AdminLabDto {
  id?: number;
  number?: number;
  topic?: string;
  description?: string;
  maxGrade?: number;
  tasks?: AdminTaskDto[];
}