// Front/src/api/generated/model/userResponseDto.ts
/**
 * Generated by orval v8.11.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { SubmissionDto } from './submissionDto';

export interface UserResponseDto {
  id?: number;
  username?: string;
  role?: string;
  firstName?: string;
  lastName?: string;
  middleName?: string;
  submissions?: SubmissionDto[];
}